preliminary changes
