+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/psi/PsiHelperFactory.kt
@@ -1,11 +1,8 @@
-package org.jetbrains.research.testspark.helpers.psi
+package org.jetbrains.research.testspark.core.psi
 
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiJavaFile
-import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.research.testspark.helpers.psi.java.JavaPsiHelper
-import org.jetbrains.research.testspark.helpers.psi.kotlin.KotlinPsiHelper
 
 object PsiHelperFactory {
     fun getPsiHelper(psiFile: PsiFile?): PsiHelper? {
