+++ java/src/org/jetbrains/research/testspark/java/JavaClassBuilderHelper.kt
