+++ kotlin/resources/org.jetbrains.research.testgenie.kotlin.xml
@@ -1,0 +1,5 @@
+<idea-plugin package="org.jetbrains.research.testgenie.kotlin">
+    <dependencies>
+        <plugin id="com.intellij.modules.java"/>
+    </dependencies>
+</idea-plugin>
