+++ kotlin/src/org/jetbrains/research/testspark/kotlin/KotlinPsiClassWrapper.kt
