+++ src/main/resources/META-INF/plugin.xml
@@ -145,8 +145,8 @@
     </change-notes>
 
     <depends>com.intellij.modules.platform</depends>
-    <depends>com.intellij.modules.java</depends>
-    <depends>org.jetbrains.kotlin</depends>
+<!--    <depends>com.intellij.modules.java</depends>-->
+<!--    <depends>org.jetbrains.kotlin</depends>-->
 
     <extensions defaultExtensionNs="com.intellij">
         <appStarter
